requirements.txt
python-socketio
python-gnupg
vincenty
aiohttp
polyline
aiohttp-cors